update dependencies, code cleanup
