+++ components/Footer.tsx
@@ -1,12 +1,11 @@
-import { type ComponentType, createElement, useContext } from 'react';
-import { StyleSheet, View, type ViewStyle } from 'react-native';
+import { useContext } from 'react';
+import { StyleSheet, View } from 'react-native';
 
 import { A, P, colors, darkColors, useLayout } from '~/common/styleguide';
 import CustomAppearanceContext from '~/context/CustomAppearanceContext';
 
 import ContentContainer from './ContentContainer';
 import {
-  type IconProps,
   Logo,
   PlatformAndroid,
   PlatformIOS,
@@ -16,38 +15,8 @@ import {
   PlatformWeb,
   PlatformWindows,
 } from './Icons';
+import Platform from './PlatformTile';
 import VercelBanner from './VercelBanner';
-
-type PlatformProps = {
-  name: string;
-  pkgName: string;
-  url: string;
-  Icon: ComponentType<IconProps>;
-  style?: ViewStyle;
-};
-
-function Platform({ name, pkgName, url, Icon, style }: PlatformProps) {
-  const { isDark } = useContext(CustomAppearanceContext);
-
-  const packageHoverStyle = {
-    backgroundColor: isDark ? darkColors.background : colors.gray2,
-    borderRadius: 8,
-  };
-
-  return (
-    <A href={url} style={styles.itemLink} hoverStyle={packageHoverStyle}>
-      <View style={[styles.platformItem, style]}>
-        {createElement(Icon, {
-          fill: isDark ? darkColors.pewter : colors.gray5,
-          width: 32,
-          height: 32,
-        })}
-        <P style={styles.platformName}>{name}</P>
-        <P style={styles.platformPackageName}>{pkgName}</P>
-      </View>
-    </A>
-  );
-}
 
 export default function Footer() {
   const { isDark } = useContext(CustomAppearanceContext);
@@ -170,24 +139,6 @@ const styles = StyleSheet.create({
     maxWidth: 960,
     marginHorizontal: 'auto',
   },
-  platformItem: {
-    minWidth: 160,
-    paddingHorizontal: 8,
-    paddingVertical: 16,
-    borderRadius: 8,
-    alignItems: 'center',
-  },
-  platformName: {
-    marginTop: 12,
-  },
-  platformPackageName: {
-    fontSize: 12,
-    fontFamily: 'monospace',
-    borderRadius: 4,
-    paddingHorizontal: 8,
-    lineHeight: 22,
-    marginTop: 2,
-  },
   bannerContainer: {
     alignSelf: 'center',
   },
@@ -204,5 +155,4 @@ const styles = StyleSheet.create({
     paddingTop: 48,
     paddingBottom: 32,
   },
-  itemLink: { backgroundColor: 'none', borderWidth: 1, borderColor: 'transparent' },
 });
