--- components/Library/Thumbnail.native.tsx
@@ -1,3 +1,0 @@
-export default function NotImplemented() {
-  return null;
-}
