+++ components/Library/Thumbnail.tsx
