+++ package.json
@@ -18,18 +18,18 @@
     "precommit": "simple-git-hooks && lint-staged"
   },
   "dependencies": {
-    "@expo/html-elements": "^0.13.7",
+    "@expo/html-elements": "^0.13.8",
     "@expo/match-media": "^0.4.0",
     "@m2d/react-markdown": "^1.0.0",
     "@radix-ui/react-hover-card": "^1.1.15",
     "@radix-ui/react-tooltip": "^1.2.8",
     "@react-native-async-storage/async-storage": "^2.2.0",
     "@react-native-picker/picker": "^2.11.4",
-    "@sentry/react": "^10.28.0",
+    "@sentry/react": "^10.29.0",
     "crypto-js": "^4.2.0",
     "es-toolkit": "^1.42.0",
-    "expo": "54.0.25",
-    "expo-font": "^14.0.9",
+    "expo": "54.0.27",
+    "expo-font": "^14.0.10",
     "next": "^16.0.7",
     "node-emoji": "^2.2.0",
     "react": "19.2.1",
